+++ src/app/(auth)/admin/autospam/index.tsx
@@ -11,7 +11,7 @@ import { ActivityIndicator, Alert, FlatList, StyleSheet } from 'react-native'
 import FastImage from 'react-native-fast-image'
 import { PressableOpacity } from 'react-native-pressable-opacity'
 import { SafeAreaView } from 'react-native-safe-area-context'
-import { PixelfedBottomSheetModal } from 'src/components/BottomSheets'
+import { PixelfedBottomSheetModal } from 'src/components/common/BottomSheets'
 import ReadMore from 'src/components/common/ReadMore'
 import UserAvatar from 'src/components/common/UserAvatar'
 import { getAutospamReports, postAutospamHandle } from 'src/lib/api'
