+++ src/app/(auth)/admin/reports/index.tsx
@@ -11,7 +11,7 @@ import { ActivityIndicator, Alert, FlatList, StyleSheet } from 'react-native'
 import FastImage from 'react-native-fast-image'
 import { PressableOpacity } from 'react-native-pressable-opacity'
 import { SafeAreaView } from 'react-native-safe-area-context'
-import { PixelfedBottomSheetModal } from 'src/components/BottomSheets'
+import { PixelfedBottomSheetModal } from 'src/components/common/BottomSheets'
 import ReadMore from 'src/components/common/ReadMore'
 import { getModReports, postReportHandle, postUserHandle } from 'src/lib/api'
 import { _timeAgo, enforceLen, htmlToTextWithLineBreaks } from 'src/utils'
