+++ src/app/(auth)/profile/[id].tsx
@@ -43,7 +43,7 @@ import { Storage } from 'src/state/cache'
 import { Button, Separator, Text, View, YStack, ZStack } from 'tamagui'
 
 import type { BottomSheetBackdropProps } from '@gorhom/bottom-sheet'
-import { PixelfedBottomSheetModal } from 'src/components/BottomSheets'
+import { PixelfedBottomSheetModal } from 'src/components/common/BottomSheets'
 
 const SCREEN_WIDTH = Dimensions.get('screen').width
